refactor: Transform Sidebar to a wrapper component
