+++ src/content-tags-drawer/index.scss
@@ -1,2 +1,2 @@
 @import "content-tags-drawer/TagBubble";
-@import "content-tags-drawer/tags-sidebar/TagsSidebar";
+@import "content-tags-drawer/tags-sidebar-controls/TagsSidebarControls";
