+++ src/content-tags-drawer/tags-sidebar-controls/index.jsx
@@ -1,0 +1,13 @@
+import TagsSidebarHeader from './TagsSidebarHeader';
+import TagsSidebarBody from './TagsSidebarBody';
+
+const TagsSidebarControls = () => (
+  <>
+    <TagsSidebarHeader />
+    <TagsSidebarBody />
+  </>
+);
+
+TagsSidebarControls.propTypes = {};
+
+export default TagsSidebarControls;
