--- src/content-tags-drawer/tags-sidebar/index.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export { default as TagsSidebarBody } from './TagsSidebarBody';
